+++ packages/wepy-cli/bin/wepy-build.js
@@ -1,0 +1,11 @@
+const compile = require('../core/compile');
+
+exports = module.exports = (program) => {
+
+  let compilation = compile(program);
+
+  debugger;
+
+  compilation.start();
+
+}