+++ packages/wepy2/index.js
@@ -1,0 +1,4 @@
+import wepy from './weapp/wepy';
+
+
+export default wepy;