+++ packages/wepy2/weapp/init/index.js
@@ -1,0 +1,5 @@
+export * from './data';
+export * from './computed';
+export * from './lifecycle';
+export * from './methods';
+export * from './watch';