+++ packages/wepy2/weapp/init/watch.js
@@ -1,0 +1,5 @@
+export function initWatch (vm, watch) {
+  Object.keys(watch).forEach(key => {
+    vm.$watch(key, watch[key]);
+  });
+}