+++ packages/wepy2/weapp/wepy.js
@@ -1,0 +1,11 @@
+import page from './page';
+import app from './app';
+import component from './component';
+import $global from './global';
+
+export default {
+  component: component,
+  page: page,
+  app: app,
+  global: $global
+}